Add handler function and split out package
