+++ packages/handler/README.md
@@ -1,0 +1,14 @@
+# `@nfte/handler`
+
+A function to allow you to self host the api portion of NFTE in your own infrastructure
+https://nfte.app
+
+---
+
+## Documentation
+
+For full documentation, visit https://nfte.app
+
+## Authors
+
+- Sam Mason de Caires
