+++ packages/handler/src/knownContracts/index.js
@@ -1,0 +1,7 @@
+import ethBlockArt from "./ethBlockArt"
+import foundation from "./foundation"
+import rarible from "./rarible"
+import superRare from "./superrare"
+import zora from "./zora"
+
+export default [ethBlockArt, foundation, rarible, superRare, zora]
