+++ packages/handler/src/utils/getMimeType.js
