+++ packages/react/tsconfig.json
@@ -5,7 +5,9 @@
     "target": "ESNext",
     "esModuleInterop": true,
     "moduleResolution": "node",
+    "declaration": true,
     "jsx": "react",
     "baseUrl": "."
-  }
+  },
+  "include": ["src"]
 }
