+++ packages/site/pages/api/embed.js.js
@@ -1,7 +1,8 @@
 import { renderToStaticMarkup } from "react-dom/server"
-import getNFTData from "@utils/getNFTData"
+import { ethers } from "ethers"
+import { NFTData } from "@nfte/handler"
 import allowCORS from "@utils/allowCORS"
-import { NFTE, css } from "@nfte/react"
+import { NFTE, css, Embed } from "@nfte/react"
 
 function embedScript(markup, css) {
   return /*javascript*/ `
@@ -20,6 +21,14 @@ function embedScript(markup, css) {
 `
 }
 
+const nftData = new NFTData({
+  provider: new ethers.providers.CloudflareProvider(),
+  historicalProvider: new ethers.providers.InfuraProvider(
+    null,
+    process.env.INFURA_PROJECT_ID
+  ),
+})
+
 export default allowCORS(async function handler(req, res) {
   const { contract, tokenId, darkMode } = req.query
 
@@ -30,9 +39,12 @@ export default allowCORS(async function handler(req, res) {
   }
 
   try {
-    const data = await getNFTData({ contract, tokenId })
+    const data = await nftData.getData({ contract, tokenId })
+
     const html = renderToStaticMarkup(
-      <NFTE initialData={data} darkMode={darkMode} />
+      <NFTE initialData={data} darkMode={darkMode}>
+        {(props) => <Embed {...props} />}
+      </NFTE>
     )
 
     res.setHeader("Content-Type", "text/javascript; charset=utf-8")
@@ -43,6 +55,7 @@ export default allowCORS(async function handler(req, res) {
     )
     res.status(200).send(embedScript(html, css))
   } catch (error) {
+    console.log(error)
     res.status(400).json({ error: "Cannot retrieve nft data" })
   }
 })
