+++ packages/site/pages/api/nft-data.js
@@ -1,5 +1,14 @@
-import getNFTData from "@utils/getNFTData"
+import { ethers } from "ethers"
+import { NFTData } from "@nfte/handler"
 import allowCORS from "@utils/allowCORS"
+
+const provider = new ethers.providers.CloudflareProvider()
+const historicalProvider = new ethers.providers.InfuraProvider(
+  null,
+  process.env.INFURA_PROJECT_ID
+)
+
+const nftData = new NFTData({ provider, historicalProvider })
 
 export default allowCORS(async function handler(req, res) {
   const { contract, tokenId } = req.query
@@ -10,7 +19,7 @@ export default allowCORS(async function handler(req, res) {
       .json({ error: "Missing contract or tokendId parameters" })
   }
   try {
-    const data = await getNFTData({ contract, tokenId })
+    const data = await nftData.getData({ contract, tokenId })
     // Cache for 30 mins
     res.setHeader(
       "Cache-Control",
