+++ packages/site/pages/index.js
@@ -234,9 +234,7 @@ export default function Home() {
             contract={debouncedContract}
             tokenId={debouncedTokenId}
             style={{ marginLeft: "auto", marginRight: "auto" }}
-          >
-            {(props) => <Embed {...props} />}
-          </NFTE>
+          />
         </Box>
 
         <Box css={{ fontSize: "@3", fontWeight: 700, mb: "@2" }}>
